refactor: propertly export KineticSdk and other classes
